### 2025-01-19 #267 @editorialbot
[REVIEW]: Educational resources to accelerate the analysis of remote sensing data using cloud resources with Xarray
**Submitting author:** @e-marshall (<a href="http://orcid.org/0000-0001-6348-977X">Emma Marshall</a>)
**Repository:** https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
