### 2025-01-20 #267 @kls2177
Hello @mikemorris12, @r-ford, @dcamron,

  Thank you again for agreeing to review. We are now ready to go!

  Please complete the above checklists and then provide your specific review comments as issues to the submitted repo. [Here is an example of a nicely organized review](https://github.com/openjournals/jose-reviews/issues/90#issuecomment-720563453).

   If you have any questions about the process, please refer to the [JOSE docs](https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html). If you don't find what you are looking for in these docs, feel free to connect with me.

Thanks!

Karen
