### 2025-01-20 #267 @r-ford
It looks like there is a [typo in the .bib file](https://github.com/e-marshall/JOSE_tutorials_submission/blob/69ba784c30d7aafda01b8affa89661c00a82dbff/paper.bib#L26C28-L27C22) preventing the pdf from generating.
