### 2025-01-21 #267 @r-ford
Guess that wasn't the issue! [Based on the error](https://github.com/openjournals/jose-papers/actions/runs/12886456675/job/35927080252#step:3:75), my guess is that there is an issue with the affiliations. Can [affiliations include links](https://github.com/e-marshall/JOSE_tutorials_submission/blob/main/paper.md?plain=1#L16)?
