### 2025-01-21 #267 @r-ford
> thank you [@r-ford](https://github.com/r-ford) ! Sorry about that! I'm not sure if i'm allowed to engage with editorial bot or not, do you know if its okay if I run 'generate pdf' to check ?

I'm new to the JOSE review process, [but it seems fine in the documentation](https://openjournals.readthedocs.io/en/jose/editorial_bot.html#author-and-reviewers-commands).
