### 2025-01-21 #267 @mikemorris12
Hi @e-marshall, I'm looking forward to reviewing your work! Since each attempt to generate the PDF adds new messages to this thread (and we're all emailed each time), I suggest debugging the PDF generation using the tips[ on this page of the JOSS docs](https://joss.readthedocs.io/en/latest/paper.html#checking-that-your-paper-compiles). I've previously published in JOSE and I used the GitHub Action tool described in the linked page to auto-compile a sample PDF upon each update to the GitHub repository.
