### 2025-01-21 #267 @e-marshall
Hey @mikemorris12 , thank you so much! I'm really sorry for all the spamming, first time trying to figure a lot of these things out! I was just looking into running this locally. i'll update when I have done that and resolved issues
