### 2025-01-21 #267 @e-marshall
the issues should be fixed now, thank you @mikemorris12 for pointing me toward the gh action, apologies to everyone for the emails and thank you so much for taking the time to review!
