### 2025-01-24 #267 @mikemorris12
Hi @e-marshall, thanks for submitting these resources to JOSE. I think each of them provide useful information and teach practical programming skills for working with climate data. The purpose of each of the two guidebooks is clear, and the manuscript contains all of the necessary information about the submission.

However, I have several comments about the way the software is archived, the styles of writing and code, and some bugs I found when testing the code. I've got a few general comments here, as well as issues I've opened in each of the GitHub repositories. These comments must be addressed before the submission can be accepted.

**Overall Comments**

-	All the submission materials should be in a single repository, instead of 3 separate ones. The submission repository doesn’t contain any of the code for either tutorial. This will make version tracking and archiving more straightforward. @kls2177, please correct me if this isn't required, but it seems to violate the "Repository" checklist item.
-	The manuscript is too long. It is well over the requested length of 1000 words.
-  Some of the citations in the paper are missing DOIs (e.g. Kluyver et al. 2016). Double check that you can find the DOI for each cited source.
- The repositories are missing licenses and version information.

The issues with specific comments about each tutorial and section are linked below:

**ITS_LIVE Tutorial**
- [General Comments](https://github.com/e-marshall/itslive/issues/39)
- [Comments about _Software and Data_ section.](https://github.com/e-marshall/itslive/issues/40)
- [Comments about _Individual Glacier Data Inspection_.](https://github.com/e-marshall/itslive/issues/41)
- [Comments about _Individual Glacier Surface Velocity Analysis_.](https://github.com/e-marshall/itslive/issues/42)
- [Comments about _Region-Scale Glacier Analysis_.](https://github.com/e-marshall/itslive/issues/43)

**Sentinel-1 RTC Tutorial**
- [General Comments](https://github.com/e-marshall/sentinel1_rtc/issues/25)
- [Comments about _Software and Data_.](https://github.com/e-marshall/sentinel1_rtc/issues/26)
- [Comments about _Data Cleaning and Organization_.](https://github.com/e-marshall/sentinel1_rtc/issues/27)
- [Comments about _ASF GDAL VRT Approach_.](https://github.com/e-marshall/sentinel1_rtc/issues/28)
- [Comments about _ASF-processed RTC Data Inspection_.](https://github.com/e-marshall/sentinel1_rtc/issues/29)
- [Comments about _RTC Dataset Comparison_.](https://github.com/e-marshall/sentinel1_rtc/issues/30)
- [Comments about _Applied example: RTC time series analysis_.](https://github.com/e-marshall/sentinel1_rtc/issues/31)
- [Comments about Appendices.](https://github.com/e-marshall/sentinel1_rtc/issues/32)
