### 2025-01-25 #267 @kls2177
Hi @mikemorris12, this is a good question about the repository materials. This is the first time I have seen multiple repositories, so I will follow-up.

In addition, multiple jupyterbooks is unusual as well. Combining the tutorials into one jupyterbook with an overarching narrative would be helpful for learners and educators.
