### 2025-01-26 #267 @e-marshall
thank you @mikemorris12 ! I'm working on addressing your comments and I'll update this thread once I've finished. I'll also work on combining the tutorials into a single book and repo
