### 2025-01-26 e-marshall/cloud-open-source-geospatial-datacube-workflows#12 @r-ford
RF JOSE review: general + paper
- General
  - I recommend putting links to the Jupyter Books in the “About” section of the repos (that’s where I look first for links).
- Paper
  - The title is repeated as the first section, with all remaining headers being formatted as subsections. I don’t think this is intended.
  - I think the Statement of Need is more verbose than it needs to be. If words need to be cut, I think this is a good place to start.
