### 2025-01-26 e-marshall/cloud-open-source-geospatial-datacube-workflows#13 @r-ford
RF JOSE review: ITS_LIVE
- I notice that there is a level of domain knowledge that is expected, so I recommend adding some text (probably in the Introduction) about any prerequisites.
- Across all sections, I recommend removing any commented out code that isn’t referenced in the text.
- Since I have no experience with remote sensing or glaciers, I will not be commenting much on the scientific content of these notebooks.
- Section 1: Software and Data
  - I do not think it is necessary to show the block of imports.
  - Instructions on running locally are vague. I recommend full instructions including cloning the repo ([like this](https://projectpythia.org/cmip6-cookbook/README.html#running-on-your-own-machine)).
  - I’m unable to solve the conda environment specified in `environment-unpinned.yml`, so I am using the `environment.yaml` file in the [binder folder](https://github.com/e-marshall/itslive/tree/main/binder), which seems to work.
  - I do not see any “lock” files in the repository.
- Section 2: Accessing cloud-hosted ITS_LIVE data
  - On Binder, I’m getting the error “No module named ‘watermark’” (not sure what this is).
  - Using `vmax` is referenced but not used in the plot.
  - Overall, I think this notebook is successful at teaching the reader how to access the data, but I think it would be improved by adding a few more example locations, instead of accessing the same chunk of data twice.
- Section 3: Individual glacier data inspection
  - When running locally, the output from `.explore()` is not really working, since the map never completely loads for me and it starts zoomed out.
  - Can you give a little background on why the different sensors give such different results?
- Section 4: Individual glacier surface velocity analysis
  - There seems to be an issue with “sample_glacier_raster” in the [current form of the book](https://e-marshall.github.io/itslive/ind_glacier_data_analysis.html).
- Section 5: Region-scale glacier analysis
