### 2025-01-26 e-marshall/cloud-open-source-geospatial-datacube-workflows#14 @r-ford
RF JOSE review: Sentinel-1 RTC
- ASF GDAL VRT approach
  - The ~50 GB download required to run this notebook is a substantial barrier. As noted in https://github.com/e-marshall/sentinel1_rtc/issues/25, one solution is to take a subset of the data. This subset might be small enough to add to the repository.
- ASF-processed RTC data inspection
- Microsoft Planetary Computer Sentinel-1 RTC Imagery
- RTC dataset comparison
  - Can you add some text to help the reader interpret the various plots?
- Applied example: RTC time series analysis
  - Same as previous comment
