### 2025-01-26 #267 @r-ford
Thanks @e-marshall for taking the time to develop these resources. I overall found the two tutorials practical and well-motivated. My main concern is about the reproducibility of the environment/code, which is essential if the goal is for these tutorials to be usable by other instructors. Since it seems like there will be some restructuring of the repositories, I will leave my specific comments as issues in the submitted repository instead of the two content repositories.

### Comments on the general checks
- Repository: There are two tutorials, which are linked in the submitted repository. Is there any issue with this? (Already discussed above)
- License: The tutorial repositories do not have licenses.
- Version: No version is given in the submission. Having two repositories for the content may cause issues with versioning.
- Authorship: Both content repositories have visible contributions from Emma Marshall, Scott Henderson, and Deepak Cherian, while Emma Marshall and Jessica Scheick have contributed to the submitted repository (the paper).

### Specific comments
- https://github.com/e-marshall/JOSE_tutorials_submission/issues/12
- https://github.com/e-marshall/JOSE_tutorials_submission/issues/13
- https://github.com/e-marshall/JOSE_tutorials_submission/issues/14
