### 2025-01-27 #267 @kls2177
> thank you [@mikemorris12](https://github.com/mikemorris12) ! I'm working on addressing your comments and I'll update this thread once I've finished. I'll also work on combining the tutorials into a single book and repo

Hi @e-marshall, I have not received any clarification yet on multiple repos, etc., so if you are able to combine them, this is likely the best path forward. Thanks!
