### 2025-02-12 #267 @e-marshall
Hi @kls2177 , thank you, and thank you @mikemorris12  and @r-ford  for your very helpful reviews and feedback! We are currently working combining the tutorials into one book and adding material to better motivate the examples contained in the tutorials and include relevant domain knowledge.
