### 2025-04-01 #267 @e-marshall
Hi all, thank you so much for the opportunity to revise this material and for your feedback and patience while we worked on it. Below is a list of changes that have been made in response to reviewer comments.

ETA: In addition to the list below, I've commented on individual issues across the old repos and linked to relevant sections of the updated book in case that is easier for review.

### General book

- The individual tutorials are now combined into one cohesive resource, which can be found [here](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/introduction.html).
    - The original scope of the tutorials was more focused on the individual notebooks within each tutorial being ‘standalone resources’ this led to a lot of duplicated material across notebooks. The feedback about this was much appreciated and we’ve streamlined the content to remove duplicated material from the notebooks. There is now relevant material page in the background section that is linked individual notebooks.
- The book is now divided into 5 'parts': 1. Introduction, 2. Background, 3. ITS_LIVE tutorial, 4. Sentinel-1 RTC tutorial, 5. Conclusion
    - All notebooks have numbers indicating the 'part' of the book and the notebook within each part
    - File names match page titles + numbers
    - Individual notebooks are divided into alphanumeric sections to reflect different concepts
- Software imports at the top of each notebook were kept in for completeness but cell metadata tags added so that they default to ‘hidden’
- 'Other resources' sections that were in individual notebooks are now centralized in a single page at the end of the [book](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/endmatter/other_resources.html)
- Throughout the notebooks, plots are cleaned up to have rotated ticklabels and cleaner labels and titles
- We broke up some of the notebooks to be shorter and have more logical content/concept breaks
- Where possible, e.g. ITS_LIVE tutorial, [nb 5](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/itslive/nbs/5_exploratory_data_analysis_group.html#write-vector-cube-to-disk), we provide a file so that expensive calculations are not required
- In response to feedback about domain knowledge needed for examples and purpose of notebooks not being well-developed, we added [Introduction](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/intro/getting_started.html) and [Background](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/context_motivation.html) sections to contextualize and motivate the workflows demonstrated in the book. We also added a [Conclusion](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/conclusion/wrapping_up.html) section to summarize content from the tutorials and synthesize take away messages
- Scripts for each tutorial ([itslive_tools.py](https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/blob/main/book/itslive/nbs/itslive_tools.py) and [s1_tools.py](https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/blob/main/book/sentinel1/nbs/s1_tools.py)) are in the same directories as the notebooks for that tutorial
- We updated the learning goals for all notebooks and categorized them as 'concepts' and 'techniques'
- We shortened the paper manuscript
- Across all notebooks, cells that previously took a long time to run should be greatly improved. This is due to a combination of software improvements and writing intermediate files to disk so that they can be read for later re-use rather than recreated.
- The current version of the book is released as [v0.1.0](https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/releases/tag/v0.1.0)

### Reproducibility

- We added more detailed [instructions for running the tutorial materials locally](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/5_software.html#running-tutorial-materials-locally) including fixed lock files and instructions for creating virtual environments to run the tutorials using conda/mamba and pixi
- We also added [instructions for running the material on a hosted jupyterhub](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/5_software.html#running-tutorial-materials-on-a-hosted-jupyterhub)
- We ran into issues getting binder resources to spin-up correctly for the ITS_LIVE tutorial so we have removed this option from the tutorial
- Rather than require users to download the full time series for the Sentinel-1 tutorial that is quite large, we implemented [two options](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/s1_intro.html#different-ways-to-use-this-tutorial): using the full time series and a subset of the time series. the tutorial introduction page has detailed instructions for downloading both datasets. the relevant notebooks have instructions for working with the 'full' or subset time series and explanation in the text (which is based on the full time series) when the results will differ if working with the subset instead of the full time series.

###  ITS_LIVE ice velocity data tutorial

- We added background on the dataset and its scientific applications to the background [section](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/4_tutorial_data.html#inter-mission-time-series-of-land-ice-velocity-and-elevation-its-live).
- We broke up the notebooks so that they are shorter with better-defined scopes and more logical breaks
- We added an additional example to the data access notebook
- We added more background and linked to discussion in the dataset documentation about why the velocity results differ by sensor (The text block immediately above the section linked [here](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/itslive/nbs/4_exploratory_data_analysis_single.html#groupby-approach))
- Fixed `geopandas.explore()` examples to default to zoomed-in view of data
- We augmented the notebooks in this section to take advantage of recent developments in the Xarray software ecosystem such as [DataTree](https://docs.xarray.dev/en/stable/generated/xarray.DataTree.html) for collections of `xr.Dataset` objects and [Xvec](https://xvec.readthedocs.io/en/stable/) for vector data cubes.

### Sentinel-1 RTC tutorial

- “Much of the code in the notebooks is data cleaning for these specific datasets. It would be nice to highlight the more widely applicable skills the guidebook teaches within the text, to complement the bullets at the start of each section.” (https://github.com/e-marshall/sentinel1_rtc/issues/25)
    - We have added context and motivation regarding these points in the [Background section](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/data_cubes.html) and re-organized this tutorial so that there is a [notebook](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/2_wrangle_metadata.html) focusing on ‘metadata wrangling’ with explanation about the code examples in the context of organizing data for analysis.
    - We also added more background and detail on SAR principles and visualizing/interpreting SAR data both in the [background](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/4_tutorial_data.html#sentinel-1-radiometric-terrain-corrected-rtc-imagery) and throughout the notebooks, in particular [notebook 3](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/3_asf_exploratory_analysis.html).
- Where there are links to stack overflow or other references, we added an explanation of them in the main text (https://github.com/e-marshall/sentinel1_rtc/issues/25)
- “Many points made in this guide are from the point of view of the author’s learning experience. I think it would be helpful to frame things as to how they can be useful for the reader.”
    - Thank you for this suggestion, it was very valuable. When the tutorials were initially developed the idea was for them to be from a learner’s perspective but we have reworked them to be more from a ‘teaching’ perspective, i.e. code that may lead to an error now has text introducing and explaining the error rather than just ‘encountering’ it and there is much more structure and text to give the reader an idea of what they will see in each section.
- Feedback from multiple reviewers on size of data used in Sentinel-1 tutorial (https://github.com/e-marshall/sentinel1_rtc/issues/25, https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/issues/14)
    - We have added an option to use a subset of the time series (5 scenes, ~2.5GB). We decided to keep the full time series in order to demonstrate many of the data cleaning operations that arise from the time series, but we stress that it is not required for users to download the full time series. We added additional detail and instructions for both options on the [Introduction](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/s1_intro.html) page. For Notebook 5 where the 2 RTC datasets are compared, an intermediate object written to disk in an earlier notebook is used. This is provided in the github repo (49 MB) so that it does not need to be downloaded separately. We’ve tried to include clear instructions in the areas where these distinctions are relevant ([Intro](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/s1_intro.html), notebooks [1](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/1_read_asf_data.html),[2](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/2_wrangle_metadata.html),[3](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/3_asf_exploratory_analysis.html)). We hope this solution addresses the reviewers’ concerns and would welcome additional feedback on this approach.
- We removed the Applied example (formerly notebook 6) because we felt that most of what was shown was covered in other notebooks.
- We updated [code that uses local file paths](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/1_read_asf_data.html#build-lists-of-file-names-and-paths-needed-for-vrt-objects) so that they are relative paths from the root directory of the book
- We made predefined variables and functions so that users should be able to easily structure their own file path lists and dictionaries. ^same notebook section as linked above
- Code that extracts information from file names and paths [now uses regex expressions](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/2_wrangle_metadata.html#parse-file-name) rather then string slices
- Fixed issues of how files were sorted
- Fixed issue with dataset comparison related to unique time steps and product IDs (mentioned in https://github.com/e-marshall/sentinel1_rtc/issues/28)
- Regarding the comment on regridding one dataset to match the spatial resolution of the other (https://github.com/e-marshall/sentinel1_rtc/issues/30), I tested using xesmf for this but I think that it’s not necessary since the spatial area of interest is small enough. We added text about this and linked to the Project Pythia cookbook on regridding with Xarray
- Added more explanation and detail to the data visualization sections (in notebooks [4.3](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/3_asf_exploratory_analysis.html#f-data-visualization), [4.4](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/4_read_pc_data.html#c-visualize-data), and [4.5](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/5_comparing_s1_rtc_datasets.html#d-visualize-comparisons))
- We have also added an additional author, Rick Forster, who collaborated with @e-marshall on the Sentinel-1 RTC tutorial and offered helpful feedback, guidance and perspective.

Thank you again for the valuable feedback and perspective. We look forward to making any additional revisions that may be helpful.
