### 2025-04-01 e-marshall/cloud-open-source-geospatial-datacube-workflows#12 @e-marshall
thank you for this feedback @r-ford ! these changes have been incorporated in the current version of the book (w/ release [v0.1.0](https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/releases/tag/v0.1.0))
