### 2025-04-01 e-marshall/cloud-open-source-geospatial-datacube-workflows#13 @e-marshall
thank you @r-ford and my apologies for these issues. these are resolved in the current version of the book and addressed in this comment: https://github.com/openjournals/jose-reviews/issues/267#issuecomment-2768147544 but to summarize here as well:
- We added an introduction section to motivate the examples and discuss relevant domain knowledge and prerequisites
- Removed commented code
- We've added cell metadata to hide library import cells
- We've added more detailed [instructions](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/5_software.html) for running the tutorial materials locally
- We fixed the lock files for the repo (https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/blob/main/default_linux-64_conda_spec.txt and https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/blob/main/build_book_linux-64_conda_spec.txt)
- We've removed the binder instance for this tutorial due to difficulties getting it to work
- We added a new example of data access in[ notebook 1 ](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/itslive/nbs/1_accessing_itslive_s3_data.html)
- Added additional code to center the `gpd.explore()` plots so this hopefully should be resolved (one example [here](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/itslive/nbs/3_combining_raster_vector_data.html#crop-vector-data-to-spatial-extent-of-raster-data))
- We added more discussion on the different satellite sensors in notebook 4 (right above this [header](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/itslive/nbs/4_exploratory_data_analysis_single.html#groupby-approach))as well as more [background](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/4_tutorial_data.html#inter-mission-time-series-of-land-ice-velocity-and-elevation-its-live) on the dataset as a whole.
- Fixed issue with `sample_glacier_raster` in old version
