### 2025-04-01 e-marshall/cloud-open-source-geospatial-datacube-workflows#14 @e-marshall
thank you @r-ford ! To address these comments we have:
- added two options for data to download for this tutorial, the full time series and a subset of the time series. the subset is 5 scenes but was still > 2 GB so is stored in an updated zenodo [record](https://zenodo.org/records/15036782) with the full time series
- We have detailed instructions for how to use the tutorial notebooks with either dataset and how to download both datasets [here](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/s1_intro.html)
- We've added more [discussion](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/sentinel1/nbs/3_asf_exploratory_analysis.html) of the SAR backscatter plots to help the reader understand and interpret them, as well as discussion of SAR principles in the tutorial data [background](https://e-marshall.github.io/cloud-open-source-geospatial-datacube-workflows/background/4_tutorial_data.html#sentinel-1-radiometric-terrain-corrected-rtc-imagery) section.
- We removed the RTC time series analysis notebook in the interest of length and because we felt that most of the content was covered elsewhere throughout the notebooks
