### 2025-04-07 #267 @r-ford
Over the next few days, I plan on taking a look at the revisions in some more detail, but from a quick look, I think the organization of the material has been significantly improved. I also appreciate all of the additional background information!
