### 2025-04-08 #267 @kls2177
Thank you @e-marshall. You have put substantial work into these revisions and it is much appreciated and brings your resource better in line with JOSE's teaching and learning vision.

Thank you @r-ford for committing to review the revisions. @mikemorris12 are you also available to review the revisions by the end of April?
