### 2025-04-08 #267 @mikemorris12
Hi @kls2177, I've also taken a preliminary look at the revised guide. Thanks @e-marshall for the hard work on this major change, and I really like the new formatting and the additional information. I've started to go through the chapters in more detail, I'll definitely be finished by the end of April.
