### 2025-04-10 e-marshall/cloud-open-source-geospatial-datacube-workflows#12 @r-ford
Thanks for the changes! I'll close the three issues I opened and open a new one with my comments for the revised book.
