### 2025-04-10 e-marshall/cloud-open-source-geospatial-data-cube-workflows#57 @r-ford
RF JOSE additional comments
### General
- The section formatting does not render when the notebooks are being run in JupyterLab (see screenshot below for example).

![Image](https://github.com/user-attachments/assets/26ade9c0-b8e5-4277-9db5-05044ad6fb0d)

I see the headings in the [config file](https://e-marshall.github.io/itslive/_static/itslive_explore_img.png), but I'm not familiar with this method. Ultimately, I don't think it is a big deal, but it makes reading through the notebooks a little confusing.

### Sentinel-1 notebooks
- In notebook 1, I got the following error when trying to run `create_vrt_object()`:

```
    0...10...20...30...40...50...60...70...80...90...100 - done.
ERROR 4: Failed to open ../data/raster_data/subset_timeseries/vrt_files/s1_stack_vv.vrt to write.
0...10...20...30...40...50...60...70...80...90...100 - done.
ERROR 4: Failed to open ../data/raster_data/subset_timeseries/vrt_files/s1_stack_vh.vrt to write.
0...10...20...30...40...50...60...70...80...90...100 - done.
ERROR 4: Failed to open ../data/raster_data/subset_timeseries/vrt_files/s1_stack_ls_map.vrt to write.
```
Creating the VRT directory in the function seemed to fix it:
```
pathlib.Path(data_path, "vrt_files").mkdir(parents=True, exist_ok=True)
```

- When running notebook 4, my kernel dies at the "3) Seasonal variability" section. My guess is that this is a memory issue, but it seems to work fine on your end. If others are able to run this notebook without issue, I'm fine with just chalking it up to the fact that I'm running the notebook on a laptop with limited resources.
