### 2025-04-10 #267 @r-ford
Overall, nice work on reorganizing and expanding the material! There were a few small issues as I was working through the notebooks, which I have noted in an issue on the submission repo:
- https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/issues/57

I also see a couple citation issues in the manuscript (first two sections).

I would recommend this submission for publication in JOSE once these comments have been addressed.
