### 2025-04-10 e-marshall/cloud-open-source-geospatial-datacube-workflows#57 @e-marshall
Thank you @r-ford ! I just pushed a fix for the VRT directory issue and the citations
- Looking at the resource usage in nb4, it is around 7GB or below except for one operation in that section where it spikes to ~ 10GB, I will look at ways to reduce this and/or add a note in that section alerting users to this
- I will also go through and remove the use of substitutions for notebook headings to resolve that issue
