### 2025-04-11 e-marshall/cloud-open-source-geospatial-data-cube-workflows#58 @mikemorris12
MM JOSE Review - Chapter 2
**Section 2.2 Data cubes**
-	You explain the concept of raster data, which is good, but vector data is also an important part of the tutorials. The concept of a vector data (and how it differs from raster data) might be unfamiliar to some readers. I suggest adding a short mention of vector data to introduce this concept to readers. Some of the content in Section 3.5B could fit here.

**Section 2.5 Software and computing environment**
-	The file environment.yml is not located in the book sub-directory like the notebook states. It’s in the main repo directory.

**Section 2.6 Relevant concepts**
-	“If you’re working with a space-time dataset but are most interested in spatial analysis, having smaller chunks along the x and y dimensions will make spatial operations easier to parallelize.”

- I think you have this backwards. If you are doing reductions over space (e.g. spatial gradients, correlations, etc.) it’s better to have larger chunks over the spatial dimensions to reduce the amount of communication between workers. If you’re doing operations aggregating over time (i.e. trend analysis), then it’s best to use smaller spatial chunks and larger time chunks.
