### 2025-04-11 e-marshall/cloud-open-source-geospatial-data-cube-workflows#59 @mikemorris12
MM JOSE Review - Chapter 3
**Section 3.1 Accessing cloud-hosted ITS_LIVE data**
-	Typo: “axes ticks” should be “axis ticks”.
-	Why use xr.open_dataset(*, engine=’zarr’) instead of xr.open_zarr()? The end result is the same, so it’s not very important. Might be worth mentioning that xr.open_zarr exists.
-	Typo: “Use the get_bounds_polyon” should be “get_bounds_polygon”.

**Section 3.2 Working with larger than memory data**
-	Outline: Expand on points A1 and A2, they’re not meaningful to a novice reader.

**Section 3.4 Exploratory data analysis of a single glacie**r
-	 “Part 1” and “Part 2” are referenced a few times (e.g. “Part 2 will include a discussion of uncertainty and error estimates”), but it’s not clear which parts of the text these refer to. I think these labels might be from an older version. Check the other sections to make sure the internal references are all correct.

**Section 3.5 Exploratory data analysis of multiple glaciers**
-	Typo: “To start with, we will look only at glaciers larger in area than 5km2” Should have a superscript 2.
