### 2025-04-11 e-marshall/cloud-open-source-geospatial-data-cube-workflows#60 @mikemorris12
MM JOSE Review - Chapter 4
**4.1 Read Sentinel-1 data processed by ASF**
-	Possible typo: “Below is the output of extract_fnames() for two one scene directory”. Is the word “two” supposed to be there?
-	The function create_vrt_object fails if the directory <data_path>/vrt_files doesn’t already exist. Add a command to this function to make the directory before creating output_vrt_path.
-	The variable path_to_rtcs should be used when creating and opening the vrt datasets, so the user doesn’t need to change the paths manually if they store the data in a non-standard directory. This is also an issue in sections 4.2, 4.3, ad 4.5 when defining the paths to the vrt files.

**4.2 Wrangle metadata**
-	When calling s1tools.create_filenames_dict, I got a list index error. I  think this is because the when the directories <path_to_rtcs>/txt_files and <path_to_rtcs>/vrt_files exist, the call of extract_fnames inside of make_filepaths_lists returns an empty list for these directories, i.e. files_of_interest[0] doesn’t exist. Maybe this error only happens when path_to_rtcs isn’t in the parent directory of the notebook, but I think that some users (like myself) will want to store the data separately from the notebooks.

**4.3 Exploratory analysis of ASF S1 imagery**
-	There is limitation of the function metadata_processor related to the file path issue first raised in Section 4.1. The variable s1_asf_data shouldn’t be hard-coded to be under the same parent directory of the notebook’s directory. This variable should be an argument of metadata_processor, so the user can specify their own directory path.
-	Typo: “We subset the data cube to a smaller area to focus on a location interest to make computation more less computationally-intensive.” This should be “… to focus on a location of interest and to reduce the computational cost”.
-	There are two notes in Section D telling the reader to skip if they’re using the subset timeseries.
-	In the very last cell, the user should be given an opportunity to define their own output directory by changing one variable, sort of like how users are prompted to define path_to_rtcs.
