### 2025-04-11 #267 @mikemorris12
Hi @e-marshall, I've taken a thorough look at the revised book, and I'm impressed with the quality of the content. This is a major improvement from the last version, congratulations!

Along with the citation issues in the manuscript that @r-ford mentioned, I have a few section-specific comments I've raised as issues in the book repo. They are mainly typos, but I'm still not completely satisfied with the way that user-defined file paths are managed in Chapter 4.

- Chapter 2: https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/issues/58
- Chapter 3: https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/issues/59
- Chapter 4: https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/issues/60

Additionally, I notice that there is only one overall license in the GitHub repo. I think you may need a separate license (like the MIT license) for the code, i.e. the Python scripts. See [this comment](https://github.com/openjournals/jose-reviews/issues/243#issuecomment-2225660468) from the review of a submission I made to JOSE last year for more details on the code license.

Once these comments are addressed, I would recommend this submission for publication in JOSE.
