### 2025-04-12 e-marshall/cloud-open-source-geospatial-datacube-workflows#57 @e-marshall
https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/commit/1fa2a7e811b69261d6113d990fa9497e0dc3bdca removes substitutes to fix the header rendering issue
