### 2025-04-13 e-marshall/cloud-open-source-geospatial-datacube-workflows#59 @e-marshall
partially addressed in https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/commit/e5d55315a092c3f777d00d87be9641fd75e9f4a7
