### 2025-04-13 e-marshall/cloud-open-source-geospatial-datacube-workflows#60 @e-marshall
thanks for this feedback! I had not tested the notebooks with data stored in a different directory, i will update the functions to enable that for reading and writing
