### 2025-04-13 e-marshall/cloud-open-source-geospatial-datacube-workflows#59 @e-marshall
also addressed in https://github.com/e-marshall/cloud-open-source-geospatial-datacube-workflows/commit/bff321fe75406751f7f250a9b89973bf75baf851
