### 2025-04-14 e-marshall/cloud-open-source-geospatial-data-cube-workflows#58 @e-marshall
Oops, thanks for pointing that out!  I think you're right re: chunking. I made a mistake interpreting this section of the Parallelizing Xarray [docs](https://docs.xarray.dev/en/stable/user-guide/dask.html#using-dask-with-xarray):
> It can be helpful to choose chunk sizes based on your downstream analyses and to chunk as early as possible. Datasets with smaller chunks along the time axis, for example, can make time domain problems easier to parallelize since Dask can perform the same operation on each time chunk. If you’re working with a large dataset with chunks that make downstream analyses challenging, you may need to rechunk your data. This is an expensive operation though, so is only recommended when needed.

I've updated that section of the text, it now reads:
"If you're working with a space-time dataset but are most interested in spatial analysis, you want to have fewer, larger chunks to minimize communication among workers during the operation, however the chunks should not be so large that operations on an individual chunk cannot be completed in memory. "

I'm happy to make additional modifications to this as well.
