### 2025-04-14 e-marshall/cloud-open-source-geospatial-data-cube-workflows#60 @e-marshall
> **4.1 Read Sentinel-1 data processed by ASF**

>     * There are two notes in Section D telling the reader to skip if they’re using the subset timeseries.

This was intentional - because the rendered version of the book is built off of the full time series, I added notes alerting users to sections where their results will not match the rendered version of the book if they're using the subset time series. I've updated the text in these sections to be more clear about when plots in the notebook with `timeseries_type= subset` won't match the version with the full time series but i'm happy to make additional changes if that would help the clarity here
