### 2025-04-14 e-marshall/cloud-open-source-geospatial-data-cube-workflows#57 @e-marshall
>     * When running notebook 4, my kernel dies at the "3) Seasonal variability" section. My guess is that this is a memory issue, but it seems to work fine on your end. If others are able to run this notebook without issue, I'm fine with just chalking it up to the fact that I'm running the notebook on a laptop with limited resources.

I tried a few things like only including one data variable in the seasonal groupby operations but I haven't been able to get this notebook to execute with less than 8 GB RAM, I could add a note about this, or remove the seasonal variability section from this notebook entirely.
