### 2025-04-14 e-marshall/cloud-open-source-geospatial-data-cube-workflows#60 @e-marshall
To address comments related to flexible locations of downloaded asf data:
>     * The variable path_to_rtcs should be used when creating and opening the vrt datasets, so the user doesn’t need to change the paths manually if they store the data in a non-standard directory. This is also an issue in sections 4.2, 4.3, ad 4.5 when defining the paths to the vrt files.
>
>
> **4.2 Wrangle metadata**
>
>     * When calling s1tools.create_filenames_dict, I got a list index error. I  think this is because the when the directories <path_to_rtcs>/txt_files and <path_to_rtcs>/vrt_files exist, the call of extract_fnames inside of make_filepaths_lists returns an empty list for these directories, i.e. files_of_interest[0] doesn’t exist. Maybe this error only happens when path_to_rtcs isn’t in the parent directory of the notebook, but I think that some users (like myself) will want to store the data separately from the notebooks.
>
>
> **4.3 Exploratory analysis of ASF S1 imagery**
>
>     * There is limitation of the function metadata_processor related to the file path issue first raised in Section 4.1. The variable s1_asf_data shouldn’t be hard-coded to be under the same parent directory of the notebook’s directory. This variable should be an argument of metadata_processor, so the user can specify their own directory path.

I've updated the code in each of the relevant notebooks and `s1_tools.py` to allow users to enter their own path and no longer hard code the path in related functions. Example of this from [notebook 4.1](https://e-marshall.github.io/cloud-open-source-geospatial-data-cube-workflows/sentinel1/nbs/1_read_asf_data.html#build-lists-of-file-names-and-paths-needed-for-vrt-objects):
![Image](https://github.com/user-attachments/assets/d47925df-8e2c-415c-b739-1459148f3d2d)
