### 2025-04-14 e-marshall/cloud-open-source-geospatial-data-cube-workflows#57 @r-ford
Thanks for looking into it. I would suggest just adding a note about the RAM required instead of removing it.
