### 2025-04-14 e-marshall/cloud-open-source-geospatial-data-cube-workflows#57 @e-marshall
That sounds good to me, I've added a note about this on https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/commit/ebc44a90c3b35ac4423d5ea2923e17dd14d6d261
