### 2025-04-14 #267 @e-marshall
Hi all,
Thank you so much for the helpful comments. I've just merged a PR that addresses feedback from last week. In addition to these changes I made three other small modifications:

- The book had some occurrences of 'datacube' but mostly uses 'data cube'. I standardized to 'data cube' except where quoting a title of an external resource that uses 'datacube'. This changes the book url to https://e-marshall.github.io/cloud-open-source-geospatial-data-cube-workflows/introduction.html
- I added a [text box](https://e-marshall.github.io/cloud-open-source-geospatial-data-cube-workflows/itslive/nbs/4_exploratory_data_analysis_single.html#d-computations-along-time-dimension) to a few relevant sections (notebook 4, [notebook 5](https://e-marshall.github.io/cloud-open-source-geospatial-data-cube-workflows/itslive/nbs/5_exploratory_data_analysis_group.html#visualize-velocity-data) and the [tutorial data page](https://e-marshall.github.io/cloud-open-source-geospatial-data-cube-workflows/background/4_tutorial_data.html#inter-mission-time-series-of-land-ice-velocity-and-elevation-its-live)) of the ITS_LIVE tutorial to more explicitly address a complexity of the dataset. We believe that incorporating this complexity into workflows demonstrated in the book is beyond the scope of the tutorial but we direct the reader to a recent preprint that presents a method of doing so if they are interested.
- I noticed varying behavior of the rendering of `.explore()` plots in [notebook 5](https://e-marshall.github.io/cloud-open-source-geospatial-data-cube-workflows/itslive/nbs/5_exploratory_data_analysis_group.html#c-data-visualization) of the itslive tutorial; in jupyterlab the map sometimes defaults to a zoomed in view of the data and sometimes not. i added a folium map element these plots with specified map extent and zoom level to address this

ETA I generated the pdf below with corrected references
