### 2025-04-16 #267 @mikemorris12
Hi @e-marshall, thanks for the revisions in response to my comments. I'm happy with the edits you've made to the notebooks and code. However, there is still a citation issue in the manuscript on Line 29. I think that the question mark indicates an issue with the reference to @sudmannsBigEarthData2020. In paper.bib, the handle for this item is @Sudmanns_2020_big.
