### 2025-04-16 #267 @e-marshall
Shoot, I'm so sorry, I thought I'd gotten that one, this is fixed on https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/commit/2c2e2f69955cf5dd6e80a65558cb6058370352c3
