### 2025-04-28 #267 @kls2177
Hi @e-marshall, I apologize for the delay. I need another week to look over your revisions. In the meantime, perhaps you can revise one small thing. I noticed that in Part 2 your figures are not referenced in the text. It is best practice to reference figures in the text to help the reader connect the figure to the content. I recommend adding these. Also, the very first figure on the landing page does not have a caption. It does not need to be numbered because it's the landing page, but a caption would be helpful for the reader. Thanks!
