### 2025-04-29 #267 @e-marshall
Hi @kls2177 , thank you for this suggestion! I just merged a [PR](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/pull/73) that adds text references for figures. For the landing page image, I wasn't able to add a caption without also adding the auto-generated figure number so it currently has both, but i'll try to play around with it more later to remove that.
