### 2025-05-01 e-marshall/cloud-open-source-geospatial-data-cube-workflows#57 @e-marshall
@r-ford is it okay if I close this issue?
