### 2025-05-07 e-marshall/JOSE_tutorials_submission#74 @kls2177
Comments on Appendix
In the "Other Resources" section, the second "Xarray-specific resources" link is not working.
