### 2025-05-07 e-marshall/cloud-open-source-geospatial-data-cube-workflows#75 @kls2177
Comments on Part 3: ITS_LIVE
Part 3.1, A 1), "Data Variables", firs bullet point: this is not displayed in the notebook. Is this something the user is expected to look at on their own machine? We do see the 60 variables later in the notebook, but not here.

Part 3.1, B 2), in the second line of code where you open the data, I was expecting to see `chunks="auto"` based on the text directly above.

Part 3.1, C 1), `find_granule` function: you have `gdf=gdf=`. Should it not just be `gdf=`?

Part 3.2, A 1): do you need to show `dc_auto` twice?

Part 3.2, B 1): why does it say `dc_auto` at the end of this section?

Part 3.4 A 1), do you need to show `single_glacier_raster` twice?

Part 3.4 A 2), proportion time series plot - what do you want the reader to take away from this plot? Sometimes you provide some interpretation of plots and sometimes you don't. It would be helpful if you had at least one sentence for each plot providing the reader with some interpretation of what they are seeing.
