### 2025-05-07 #267 @kls2177
Hi @e-marshall, I am making my way through the revised version. I have forked your repo and suggested small wording changes here and there. There are still some typos throughout the notebooks, but I have not flagged them here.

Here are some other comments:

[Comments on Part 3: ITS_LIVE](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/issues/75)
[Comments on Appendix](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/issues/74)

Overall, everything is looking very good. I do not anticipate any major issues. I aim to finished my final review over the weekend.
