### 2025-05-12 e-marshall/cloud-open-source-geospatial-data-cube-workflows#74 @e-marshall
Thank you @kls2177 ! I've updated the link in the text to reflect the new location of the Xarray 'How do I...' page in https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/commit/546e040d3cb61d380c5c679c3938700b7489715d
