### 2025-05-12 #267 @e-marshall
Hi @kls2177 , thank you so much and my apologies for the delay. I just merged https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/pull/76which addresses the Comments in the above issues. I'm sorry about the typos, I thought I'd gotten them all with the precommit checker. I will go through it again today
