### 2025-05-23 #267 @e-marshall
Hi @kls2177, thank you so much! Just to clarify, should the zenodo record reflect the current version or were there changes on your fork that should be merged first?
