### 2025-06-04 #267 @e-marshall
Hi @kls2177, here is the link to the Zenodo archive of the resource: [10.5281/zenodo.15594587](10.5281/zenodo.15594587)
