### 2025-06-06 #267 @e-marshall
Sorry about this ^, I had the url's listed for the resources that didn't have specified doi's but they weren't under the 'doi' key of the bibtex entry. this is fixed on https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/pull/80
