### 2025-06-06 #267 @e-marshall
@kls2177 I fixed several small typos in PRs [77](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/pull/77) and [78](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/pull/78) . i didn't see a PR from your fork but I'm happy to update the archived version if there are additional changes you'd like me to add
