### 2025-06-09 #267 @kls2177
Hi @e-marshall, looks good to me. Just one error in paper.md in the first list of references: wagemann 2022 is missing the @ symbol, so the citation is not rendering properly. Once this is fixed, we will officially accept.
