### 2025-06-09 #267 @kls2177
@e-marshall, there still seems to be a problem with the Procida reference in the .bib file.
