### 2025-06-10 #267 @e-marshall
my mistake there, i had changed the bibtex entries from 'url' to 'doi' on the procida reference and the two others that were flagged. I've now changed them back to url, hopefully this works? I've also updated the bibtex so that all references used in the book, not just those in the paper, appear in the bibliography
