### 2025-06-10 #267 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
