### 2025-06-23 #267 @labarba
Please re-format the author field for the `pandas` reference, as it appears as "team, T. pandas development."
