### 2025-06-23 #267 @labarba
I see the version here (top comment) as "v1.0" but the last tagged release in the repository is [v0.2.1](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/releases/tag/v0.2.1)—does a new tagged release need to be issued? (The version here needs to match.)
