### 2025-06-23 #267 @labarba
[Review checklist for @r-ford](https://github.com/openjournals/jose-reviews/issues/267#issuecomment-2603176732) has several unchecked items. Is this just an oversight, and the review was completed?
@r-ford – could you check of the remaining items, for a clean Review record?
