### 2025-06-23 #267 @labarba
[Review checklist for @dcamron](https://github.com/openjournals/jose-reviews/issues/267#issuecomment-2608331345) has all items unchecked. Did this reviewer not complete a review? (If so, we should remove as reviewer here.)
