### 2025-06-23 #267 @dcamron
> [Review checklist for @dcamron](https://github.com/openjournals/jose-reviews/issues/267#issuecomment-2608331345) has all items unchecked. Did this reviewer not complete a review? (If so, we should remove as reviewer here.)

Hi, I have not completed a review, please do remove me. I apologize for the silence @e-marshall, and thank you for considering me. I hope the rest of the process goes smoothly.
