### 2025-06-23 #267 @mikemorris12
> [@mikemorris12](https://github.com/mikemorris12) – you leaved License unchecked in your review checklist: are you ready to check that off?

Apologies, I forgot to update the checklist following the latest revisions. I've checked off the License box.
