### 2025-06-23 #267 @labarba
@e-marshall  – You may add more than one license to a Zenodo archive: when the files you are uploading are under different licenses (e.g., software under the MIT license and documentation under a CC-BY license), you can declare all applicable licenses in the Zenodo record metadata.
https://help.zenodo.org/docs/deposit/describe-records/licenses/#:~:text=Mixed%20license%20uploads,done%20in%20the%20example%20below:
