### 2025-07-02 #267 @e-marshall
Hi @labarba  and @kls2177, my apologies for the delay! I've now done the following:
- updated the title and author affiliations in the paper to match the repo/zenodo record
- added a MIT license to the zenodo record
- released v1.0 to match the submission version in the top comment.

The top comment also has listed the old name of the repo (before the two tutorials initially submitted were combined into one book); is it possible/necessary to update this url to the new repo name ([https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows))?
