### 2025-07-02 #267 @labarba
Maybe it still remains to publish the changes on the Zenodo archive? I see the new version, but the title does not match the paper and the MIT license is not showing...
