### 2025-07-02 #267 @labarba
we rather wanted the opposite! The paper title should still be a descriptive title, rather than a repository name, and the Zenodo metadata should be edited to match...
