### 2025-07-02 #267 @labarba
I see where [the paper title was changed](https://github.com/e-marshall/cloud-open-source-geospatial-data-cube-workflows/commit/f27d8f796885041d9f3b12a1cb1aba69be26c93f#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6R2) to "Cloud-native geospatial data cube workflows with open-source tools" – that is fine. What I'm saying is that the Zenodo archive should match that title.
