### 2025-07-03 #267 @e-marshall
I'm sorry, I misunderstood that ! I think I've fixed it correctly now. Here is the updated [record](https://zenodo.org/records/15788188)

![Image](https://github.com/user-attachments/assets/6a063adf-eaa8-449f-9f4c-822ebe7c591d)

![Image](https://github.com/user-attachments/assets/052922a8-ad75-456f-8ace-56fc6006b7f4)
