### 2025-07-04 #267 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
